backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-genai==1.7.0
python-dotenv==1.0.1
pydantic==2.6.4